third_party/WebKit/LayoutTests/fast/dom/Element/id-in-getelement01-expected.txt

Test getElementById

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elem.getAttribute("name") is "name1"
PASS elem.getAttribute("name") is "name2"
PASS elem is null
PASS successfullyParsed is true

TEST COMPLETE
